# lib/api-zod/src/generated/types/pet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grooming CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { PetSize } from "./petSize";

export interface Pet {
  id: number;
  clientId: number;
  name: string;
  /** @nullable */
  breed?: string | null;
  size: PetSize;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}